por el cambio de la gestion de usuario se reacomodo el codigo
