--- src/app/admin-managent/components/side-menu-newuser/side-menu-newuser.component.html
@@ -1,35 +1,0 @@
-<div class="menu menu-default flex flex-wrap justify-center gap-2.5 border rounded-lg py-2">
-    <div class="menu-item">
-        
-        <a [routerLink]="'nuevo-investigador'" class="menu-link">
-            <span class="menu-icon">
-                <i class="ki-solid ki-react"></i>
-            </span>
-            <span class="text-lg">
-                Investigador
-            </span>
-        </a>
-    </div>
-    <div class="menu-separator">
-    </div>
-    <div class="menu-item">
-        <a [routerLink]="'nuevo-suplente'" routerLinkActive="btn-success" class="menu-link">
-            <span class="menu-icon">
-                <i class="ki-solid ki-users"></i>
-            </span>
-            <span class="text-lg ">
-                Suplente
-            </span>
-        </a>
-    </div>
-    <div class="menu-item">
-        <a [routerLink]="'nuevo-evaluador'" routerLinkActive="btn-success" class="menu-link">
-            <span class="menu-icon">
-                <i class="ki-solid ki-pencil"></i>
-            </span>
-            <span class="text-lg ">
-                Evaluador
-            </span>
-        </a>
-    </div>
-</div>
