--- src/app/admin-managent/components/side-menu-newuser/side-menu-newuser.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-import { RouterLink } from '@angular/router';
-
-@Component({
-  selector: 'side-menu-newuser',
-  imports: [RouterLink],
-  templateUrl: './side-menu-newuser.component.html',
-})
-export class SideMenuNewuserComponent { }
