--- src/app/admin-managent/pages/users-management/crear-usuario/crear-evaluador/crear-evaluador.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-import { RouterLink } from '@angular/router';
-
-@Component({
-  selector: 'app-crear-evaluador',
-  imports: [RouterLink],
-  templateUrl: './crear-evaluador.component.html',
-})
-export class CrearEvaluadorComponent { }
