--- src/app/admin-managent/pages/users-management/crear-usuario/crear-investigador/crear-investigador.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-import { RouterLink } from '@angular/router';
-
-@Component({
-  selector: 'app-crear-investigador',
-  imports: [RouterLink],
-  templateUrl: './crear-investigador.component.html',
-})
-export class CrearInvestigadorComponent { }
