--- src/app/admin-managent/pages/users-management/crear-usuario/crear-suplente/crear-suplente.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-import { RouterLink } from '@angular/router';
-
-@Component({
-  selector: 'app-crear-suplente',
-  imports: [RouterLink],
-  templateUrl: './crear-suplente.component.html',
-})
-export class CrearSuplenteComponent { }
