--- src/app/admin-managent/pages/users-management/modificar-usuario/modificar-usuario.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-modificar-usuario',
-  imports: [],
-  templateUrl: './modificar-usuario.component.html',
-})
-export class ModificarUsuarioComponent { }
