--- src/app/admin-managent/pages/users-management/users-list/users-list.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'users-list',
-  imports: [],
-  templateUrl: './users-list.component.html',
-})
-export class UsersListComponent { }
